+++ RayTracingWeekend/camera.h
@@ -10,7 +10,7 @@ public:
 	camera() {}
 
 	// vfov is top of bottom in degree
-	camera(const vec3& lookfrom, const vec3& lookat, const vec3& vup, float vfov, float aspect, float aperture, float focus_dist, float t0, float t1)
+	camera(const vec3& lookfrom, const vec3& lookat, const vec3& vup, double vfov, double aspect, double aperture, double focus_dist, double t0, double t1)
 	{
 		time0 = t0;
 		time1 = t1;
@@ -18,9 +18,9 @@ public:
 		// so aperture is diameter...
 		lens_radius = aperture / 2;
 
-		float theta = vfov * static_cast<float>(M_PI) / 180.0f;
-		float half_height = tan(theta / 2);
-		float half_width = aspect * half_height;
+		double theta = vfov * static_cast<double>(M_PI) / 180.0;
+		double half_height = tan(theta / 2);
+		double half_width = aspect * half_height;
 
 		origin = lookfrom;
 		w = normalize(lookfrom - lookat);
@@ -29,16 +29,16 @@ public:
 		
 		lower_left_corner = origin - half_width * focus_dist *  u - half_height * focus_dist * v - focus_dist * w;
 		
-		horizontal = 2.0f * half_width * focus_dist * u;
-		vertical = 2.0f * half_height * focus_dist * v;		
+		horizontal = 2.0 * half_width * focus_dist * u;
+		vertical = 2.0 * half_height * focus_dist * v;		
 	}
 
-	ray get_ray(float s, float t)
+	ray get_ray(double s, double t)
 	{
 		vec3 rd = lens_radius * random_in_unit_disk();
 		vec3 offset = u * rd.x + v * rd.y;
 
-		float time = time0 + uniform(timeEngine) * (time1 - time0);
+		double time = time0 + uniform(timeEngine) * (time1 - time0);
 
 		auto dir = lower_left_corner
 			+ s * horizontal
@@ -54,8 +54,8 @@ public:
 	vec3 horizontal;
 	vec3 vertical;
 	vec3 u, v, w;
-	float time0, time1;
-	float lens_radius;
+	double time0, time1;
+	double lens_radius;
 
 private:
 	vec3 random_in_unit_disk()
@@ -63,12 +63,12 @@ private:
 		vec3 p;
 		do
 		{
-			p = 2.0f * vec3(uniform(rayEngine), uniform(rayEngine), 0) - vec3(1, 1, 0);
-		} while (dot(p, p) >= 1.0f);
+			p = 2.0 * vec3(uniform(rayEngine), uniform(rayEngine), 0) - vec3(1, 1, 0);
+		} while (dot(p, p) >= 1.0);
 		return p;
 	}
 
-	std::uniform_real_distribution<float> uniform;
+	std::uniform_real_distribution<double> uniform;
 	std::minstd_rand rayEngine;
 	std::minstd_rand timeEngine;
 };
